factor out mccs
